+++ src/plugin/PluginEditor.h
@@ -74,6 +74,7 @@ private:
     juce::Slider pickTextureSlider;
     juce::Slider pickBiteSlider;
     juce::ComboBox pickStrokeBox;
+    juce::Slider strumSpeedSlider;
     juce::Slider playerFeelSlider;
     juce::Slider playerFeelRecoverySlider;
     juce::Slider palmMuteSlider;
@@ -112,7 +113,11 @@ private:
     juce::Label pickTextureLabel;
     juce::Label pickBiteLabel;
     juce::Label pickStrokeLabel;
+    juce::Label strumSpeedLabel;
     juce::Label playerFeelLabel;
+    juce::Label playerFeelBotLabel;
+    juce::Label playerFeelProLabel;
+    juce::Label playerFeelLooseLabel;
     juce::Label playerFeelRecoveryLabel;
     juce::Label playerFeelCognitiveLabel;
     juce::Label playerFeelDexterityLabel;
@@ -154,6 +159,7 @@ private:
     juce::TextButton pickTextureInfoButton;
     juce::TextButton pickBiteInfoButton;
     juce::TextButton pickStrokeInfoButton;
+    juce::TextButton strumSpeedInfoButton;
     juce::TextButton playerFeelInfoButton;
     juce::TextButton playerFeelRecoveryInfoButton;
     juce::TextButton playerFeelResetButton;
@@ -209,6 +215,7 @@ private:
     std::unique_ptr<SliderAttachment> pickTextureAttachment;
     std::unique_ptr<SliderAttachment> pickBiteAttachment;
     std::unique_ptr<ComboBoxAttachment> pickStrokeAttachment;
+    std::unique_ptr<SliderAttachment> strumSpeedAttachment;
     std::unique_ptr<SliderAttachment> playerFeelAttachment;
     std::unique_ptr<SliderAttachment> playerFeelRecoveryAttachment;
     std::unique_ptr<SliderAttachment> palmMuteAttachment;